+++ example/denoise.js
@@ -16,7 +16,7 @@ import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
 import { GUI } from 'three/examples/jsm/libs/lil-gui.module.min.js';
 import { Upscaler } from '@pmndrs/upscaler';
 import { LoaderElement } from './src/LoaderElement.js';
-import { OIDNDenoiser } from './src/OIDNDenoiser.js';
+import { OIDNDenoiser } from './src/denoise/OIDNDenoiser.js';
 import { GradientEquirectTexture } from 'three-gpu-pathtracer';
 import { WebGPUPathTracer, RenderToScreenNodeMaterial } from 'three-gpu-pathtracer/webgpu';
